refactor: Move CoilSetup to App
